+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Change `total` key to `total_items`
+
 # 0.0.1
 
 - Initial release